Move inference under swebench package
