+++ swebench/inference/make_datasets/create_instance.py
@@ -8,12 +8,8 @@ from tempfile import TemporaryDirectory
 import unidiff
 from tqdm.auto import tqdm
 
-try:
-    from tokenize_dataset import TOKENIZER_FUNCS
-    from utils import AutoContextManager, ingest_directory_contents
-except:
-    from .tokenize_dataset import TOKENIZER_FUNCS
-    from .utils import AutoContextManager, ingest_directory_contents
+from swebench.inference.make_datasets.tokenize_dataset import TOKENIZER_FUNCS
+from swebench.inference.make_datasets.utils import AutoContextManager, ingest_directory_contents
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s %(levelname)s %(message)s")
 logger = logging.getLogger(__name__)
