+++ swebench/inference/run_live.py
@@ -15,20 +15,20 @@ import re
 import time
 from datetime import datetime
 from tqdm.auto import tqdm
-from make_datasets.utils import ContextManager, string_to_bool, extract_diff, extract_minimal_patch
-from make_datasets.bm25_retrieval import (
-    make_index,
-    clone_repo,
-    search,
-    DOCUMENT_ENCODING_FUNCTIONS,
-)
-from make_datasets.create_instance import (
+from swebench.inference.make_datasets.utils import ContextManager, string_to_bool, extract_diff, extract_minimal_patch
+from swebench.inference.make_datasets.create_instance import (
     PROMPT_FUNCTIONS,
     TOKENIZER_FUNCS,
     make_code_text,
     ingest_files,
 )
-from run_api import call_chat, call_anthropic
+from swebench.inference.make_datasets.bm25_retrieval import (
+    make_index,
+    clone_repo,
+    search,
+    DOCUMENT_ENCODING_FUNCTIONS,
+)
+from swebench.inference.run_api import call_chat, call_anthropic
 import logging
 from argparse import ArgumentParser
